refactor(frontend): split box prop rendering
